LPS-53617 EditFolderAction: move logic to portlet class (no logic changes)
